requirements.txt
supervisor
python-telegram-bot
sqlalchemy
alembic
requests
tmdbv3api
apscheduler
pytz